requirements.txt
torch
numpy
pillow
imageio
opencv-python
